Move version name dialog to FSpot.UI.Dialog.
