+++ src/Clients/FSpot/FSpot.UI.Dialog/ui/VersionNameDialog.ui
